+++ components/site/src/lib.rs
@@ -39,7 +39,7 @@ use utils::net::get_available_port;
 use content::{Page, Section, populate_previous_and_next_pages, sort_pages};
 use templates::{GUTENBERG_TERA, global_fns, render_redirect_template};
 use front_matter::{SortBy, InsertAnchor};
-use taxonomies::Taxonomy;
+use taxonomies::{Taxonomy, find_taxonomies};
 use pagination::Paginator;
 
 use rayon::prelude::*;
@@ -74,8 +74,7 @@ pub struct Site {
     pub output_path: PathBuf,
     content_path: PathBuf,
     pub static_path: PathBuf,
-    pub tags: Option<Taxonomy>,
-    pub categories: Option<Taxonomy>,
+    pub taxonomies: Vec<Taxonomy>,
     /// A map of all .md files (section and pages) and their permalink
     /// We need that if there are relative links in the content that need to be resolved
     pub permalinks: HashMap<String, String>,
@@ -128,8 +127,7 @@ impl Site {
             output_path: path.join("public"),
             content_path,
             static_path,
-            tags: None,
-            categories: None,
+            taxonomies: Vec::new(),
             permalinks: HashMap::new(),
         };
 
@@ -248,7 +246,7 @@ impl Site {
         self.register_early_global_fns();
         self.render_markdown()?;
         self.populate_sections();
-        self.populate_tags_and_categories();
+        self.populate_taxonomies();
         self.register_tera_global_fns();
 
         Ok(())
@@ -302,7 +300,7 @@ impl Site {
         self.tera.register_global_function("get_section", global_fns::make_get_section(&self.sections));
         self.tera.register_global_function(
             "get_taxonomy_url",
-            global_fns::make_get_taxonomy_url(self.tags.clone(), self.categories.clone())
+            global_fns::make_get_taxonomy(self.taxonomies.clone())
         );
     }
 
@@ -409,15 +407,12 @@ impl Site {
     }
 
     /// Find all the tags and categories if it's asked in the config
-    pub fn populate_tags_and_categories(&mut self) {
-        let generate_tags_pages = self.config.generate_tags_pages;
-        let generate_categories_pages = self.config.generate_categories_pages;
-        if !generate_tags_pages && !generate_categories_pages {
+    pub fn populate_taxonomies(&mut self) {
+        if self.config.taxonomies.is_empty() {
             return;
         }
 
-        // TODO: can we pass a reference?
-        let (tags, categories) = Taxonomy::find_tags_and_categories(
+        self.taxonomies = find_taxonomies(
             &self.config,
             self.pages
                 .values()
@@ -426,12 +421,6 @@ impl Site {
                 .collect::<Vec<_>>()
                 .as_slice()
         );
-        if generate_tags_pages {
-            self.tags = Some(tags);
-        }
-        if generate_categories_pages {
-            self.categories = Some(categories);
-        }
     }
 
     /// Inject live reload script tag if in live reload mode
@@ -524,14 +513,11 @@ impl Site {
         self.render_orphan_pages()?;
         self.render_sitemap()?;
         if self.config.generate_rss {
-            self.render_rss_feed()?;
+            self.render_rss_feed(None, None)?;
         }
         self.render_404()?;
         self.render_robots()?;
-        // `render_categories` and `render_tags` will check whether the config allows
-        // them to render or not
-        self.render_categories()?;
-        self.render_tags()?;
+        self.render_taxonomies()?;
 
         if let Some(ref theme) = self.config.theme {
             let theme_path = self.base_path.join("themes").join(theme);
@@ -681,19 +667,11 @@ impl Site {
         )
     }
 
-    /// Renders all categories and the single category pages if there are some
-    pub fn render_categories(&self) -> Result<()> {
-        if let Some(ref categories) = self.categories {
-            self.render_taxonomy(categories)?;
-        }
-
-        Ok(())
-    }
-
-    /// Renders all tags and the single tag pages if there are some
-    pub fn render_tags(&self) -> Result<()> {
-        if let Some(ref tags) = self.tags {
-            self.render_taxonomy(tags)?;
+    /// Renders all taxonomies with at least one non-draft post
+    pub fn render_taxonomies(&self) -> Result<()> {
+        // TODO: make parallel?
+        for taxonomy in &self.taxonomies {
+            self.render_taxonomy(taxonomy)?;
         }
 
         Ok(())
@@ -705,8 +683,8 @@ impl Site {
         }
 
         ensure_directory_exists(&self.output_path)?;
-        let output_path = self.output_path.join(&taxonomy.get_list_name());
-        let list_output = taxonomy.render_list(&self.tera, &self.config)?;
+        let output_path = self.output_path.join(&taxonomy.kind.name);
+        let list_output = taxonomy.render_all_terms(&self.tera, &self.config)?;
         create_directory(&output_path)?;
         create_file(&output_path.join("index.html"), &self.inject_livereload(list_output))?;
 
@@ -714,12 +692,25 @@ impl Site {
             .items
             .par_iter()
             .map(|item| {
-                let single_output = taxonomy.render_single_item(item, &self.tera, &self.config)?;
-                create_directory(&output_path.join(&item.slug))?;
-                create_file(
-                    &output_path.join(&item.slug).join("index.html"),
-                    &self.inject_livereload(single_output)
-                )
+                if taxonomy.kind.rss {
+                    // TODO: can we get rid of `clone()`?
+                    self.render_rss_feed(
+                        Some(item.pages.clone()),
+                        Some(&PathBuf::from(format!("{}/{}", taxonomy.kind.name, item.slug)))
+                    )?;
+                }
+
+                if taxonomy.kind.is_paginated() {
+                    self.render_paginated(&output_path, &Paginator::from_taxonomy(&taxonomy, item))
+                } else {
+                    let single_output = taxonomy.render_term(item, &self.tera, &self.config)?;
+                    let path = output_path.join(&item.slug);
+                    create_directory(&path)?;
+                    create_file(
+                        &path.join("index.html"),
+                        &self.inject_livereload(single_output),
+                    )
+                }
             })
             .fold(|| Ok(()), Result::and)
             .reduce(|| Ok(()), Result::and)
@@ -752,31 +743,19 @@ impl Site {
         sections.sort_by(|a, b| a.permalink.cmp(&b.permalink));
         context.add("sections", &sections);
 
-        let mut categories = vec![];
-        if let Some(ref c) = self.categories {
-            let name = c.get_list_name();
-            categories.push(SitemapEntry::new(self.config.make_permalink(&name), None));
-            for item in &c.items {
-                categories.push(
-                    SitemapEntry::new(self.config.make_permalink(&format!("{}/{}", &name, item.slug)), None),
-                );
+        let mut taxonomies = vec![];
+        for taxonomy in &self.taxonomies {
+            let name = &taxonomy.kind.name;
+            let mut terms = vec![];
+            terms.push(SitemapEntry::new(self.config.make_permalink(name), None));
+            for item in &taxonomy.items {
+                terms.push(SitemapEntry::new(self.config.make_permalink(&format!("{}/{}", &name, item.slug)), None));
             }
+            terms.sort_by(|a, b| a.permalink.cmp(&b.permalink));
+            taxonomies.push(terms);
         }
-        categories.sort_by(|a, b| a.permalink.cmp(&b.permalink));
-        context.add("categories", &categories);
+        context.add("taxonomies", &taxonomies);
 
-        let mut tags = vec![];
-        if let Some(ref t) = self.tags {
-            let name = t.get_list_name();
-            tags.push(SitemapEntry::new(self.config.make_permalink(&name), None));
-            for item in &t.items {
-                tags.push(
-                    SitemapEntry::new(self.config.make_permalink(&format!("{}/{}", &name, item.slug)), None),
-                );
-            }
-        }
-        tags.sort_by(|a, b| a.permalink.cmp(&b.permalink));
-        context.add("tags", &tags);
         context.add("config", &self.config);
 
         let sitemap = &render_template("sitemap.xml", &self.tera, &context, &self.config.theme)?;
@@ -786,14 +765,20 @@ impl Site {
         Ok(())
     }
 
-    pub fn render_rss_feed(&self) -> Result<()> {
+    /// Renders a RSS feed for the given path and at the given path
+    /// If both arguments are `None`, it will render only the RSS feed for the whole
+    /// site at the root folder.
+    pub fn render_rss_feed(&self, all_pages: Option<Vec<Page>>, base_path: Option<&PathBuf>) -> Result<()> {
         ensure_directory_exists(&self.output_path)?;
 
         let mut context = Context::new();
-        let pages = self.pages.values()
+        let pages = all_pages
+            // TODO: avoid that cloned().
+            // It requires having `sort_pages` take references of Page
+            .unwrap_or_else(|| self.pages.values().cloned().collect::<Vec<_>>())
+            .into_iter()
             .filter(|p| p.meta.date.is_some() && !p.is_draft())
-            .cloned()
-            .collect::<Vec<Page>>();
+            .collect::<Vec<_>>();
 
         // Don't generate a RSS feed if none of the pages has a date
         if pages.is_empty() {
@@ -802,20 +787,32 @@ impl Site {
 
         let (sorted_pages, _) = sort_pages(pages, SortBy::Date);
         context.add("last_build_date", &sorted_pages[0].meta.date.clone().map(|d| d.to_string()));
-         // limit to the last n elements)
+         // limit to the last n elements
         context.add("pages", &sorted_pages.iter().take(self.config.rss_limit).collect::<Vec<_>>());
         context.add("config", &self.config);
 
-        let rss_feed_url = if self.config.base_url.ends_with('/') {
-            format!("{}{}", self.config.base_url, "rss.xml")
+        let rss_feed_url = if let Some(ref base) = base_path {
+            self.config.make_permalink(&base.join("rss.xml").to_string_lossy())
         } else {
-            format!("{}/{}", self.config.base_url, "rss.xml")
+            self.config.make_permalink("rss.xml")
         };
+
         context.add("feed_url", &rss_feed_url);
 
         let feed = &render_template("rss.xml", &self.tera, &context, &self.config.theme)?;
 
-        create_file(&self.output_path.join("rss.xml"), feed)?;
+        if let Some(ref base) = base_path {
+            let mut output_path = self.output_path.clone().to_path_buf();
+            for component in base.components() {
+                output_path.push(component);
+                if !output_path.exists() {
+                    create_directory(&output_path)?;
+                }
+            }
+            create_file(&output_path.join("rss.xml"), feed)?;
+        } else {
+            create_file(&self.output_path.join("rss.xml"), feed)?;
+        }
 
         Ok(())
     }
@@ -854,7 +851,7 @@ impl Site {
         }
 
         if section.meta.is_paginated() {
-            self.render_paginated(&output_path, section)?;
+            self.render_paginated(&output_path, &Paginator::from_section(&section.pages, section))?;
         } else {
             let output = section.render_html(&self.tera, &self.config)?;
             create_file(&output_path.join("index.html"), &self.inject_livereload(output))?;
@@ -894,11 +891,10 @@ impl Site {
     }
 
     /// Renders a list of pages when the section/index is wanting pagination.
-    pub fn render_paginated(&self, output_path: &Path, section: &Section) -> Result<()> {
+    pub fn render_paginated(&self, output_path: &Path, paginator: &Paginator) -> Result<()> {
         ensure_directory_exists(&self.output_path)?;
 
-        let paginator = Paginator::new(&section.pages, section);
-        let folder_path = output_path.join(&section.meta.paginate_path);
+        let folder_path = output_path.join(&paginator.paginate_path);
         create_directory(&folder_path)?;
 
         paginator
@@ -913,7 +909,7 @@ impl Site {
                     create_file(&page_path.join("index.html"), &self.inject_livereload(output))?;
                 } else {
                     create_file(&output_path.join("index.html"), &self.inject_livereload(output))?;
-                    create_file(&page_path.join("index.html"), &render_redirect_template(&section.permalink, &self.tera)?)?;
+                    create_file(&page_path.join("index.html"), &render_redirect_template(&paginator.permalink, &self.tera)?)?;
                 }
                 Ok(())
             })